[WB-1816] Rework `semanticColor` tokens to be exported as CSS variables (#2481)
## Summary:

- Tokens: Replaces `semanticColor` values with references to css variables.
Also, adds a build step to convert `semanticColor` tokens to css variables that
can be consumed via CSS imports.
- Theming: Adds a new `ThemeSwitcher` component to change themes using CSS
variables.
- Pill: Replaces `mix()` with `color-mix()` for active background colors.

Issue: WB-1816

## Test plan:

1. Navigate to `/?path=/story/packages-labeledfield--validation`.
2. Click on the `Theme` dropdown in the storybook toolbar.
3. Select a different theme from the dropdown (`Classroom`).
4. Verify that the theme changes and the colors are updated.

https://github.com/user-attachments/assets/33848f65-599c-42db-aac6-47b5b87fce76

Author: jandrade

Reviewers: jandrade, marcysutton, beaesguerra

Required Reviewers:

Approved By: marcysutton, beaesguerra

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2481
