+++ .changeset/gold-peaches-guess.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/wonder-blocks-form": patch
+---
+
+TextField: Fix outline-width on focus
