+++ .changeset/heavy-islands-tan.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/wonder-blocks-pill": patch
+---
+
+Use color-mix() for active background colors
