+++ .changeset/ten-gorillas-invent.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/wonder-blocks-icon-button": patch
+---
+
+Simplify IconButton pseudo-classes
