+++ .changeset/tidy-humans-fail.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/wonder-blocks-theming": minor
+---
+
+Add ThemeSwitcher component to change themes using CSS variables
