+++ .storybook/preview.tsx
@@ -4,9 +4,18 @@ import {Decorator} from "@storybook/react";
 import {semanticColor} from "@khanacademy/wonder-blocks-tokens";
 import {initAnnouncer} from "@khanacademy/wonder-blocks-announcer";
 import Link from "@khanacademy/wonder-blocks-link";
-import {ThemeSwitcherContext} from "@khanacademy/wonder-blocks-theming";
+import {
+    ThemeSwitcherContext,
+    ThemeSwitcher,
+} from "@khanacademy/wonder-blocks-theming";
 import {RenderStateRoot} from "../packages/wonder-blocks-core/src";
 import {Preview} from "@storybook/react";
+
+// Import the Wonder Blocks CSS variables
+// NOTE: External consumers should import the CSS variables from the
+// wonder-blocks-tokens package directly.
+// e.g. import "@khanacademy/wonder-blocks-tokens/styles.css";
+import "../node_modules/@khanacademy/wonder-blocks-tokens/dist/css/index.css";
 
 /**
  * WB Official breakpoints
@@ -104,14 +113,18 @@ const withThemeSwitcher: Decorator = (
         return (
             <RenderStateRoot>
                 <ThemeSwitcherContext.Provider value={theme}>
-                    <Story />
+                    <ThemeSwitcher theme={theme}>
+                        <Story />
+                    </ThemeSwitcher>
                 </ThemeSwitcherContext.Provider>
             </RenderStateRoot>
         );
     }
     return (
         <ThemeSwitcherContext.Provider value={theme}>
-            <Story />
+            <ThemeSwitcher theme={theme}>
+                <Story />
+            </ThemeSwitcher>
         </ThemeSwitcherContext.Provider>
     );
 };
@@ -195,7 +208,7 @@ const preview: Preview = {
                     },
                     {
                         value: "khanmigo",
-                        icon: "circle",
+                        icon: "comment",
                         title: "Khanmigo",
                     },
                 ],
