+++ package.json
@@ -9,7 +9,8 @@
   },
   "scripts": {
     "alex": "alex packages/",
-    "build": "rollup -c ./build-settings/rollup.config.mjs",
+    "dev": "pnpm -r dev",
+    "build": "rollup -c ./build-settings/rollup.config.mjs && pnpm -r build",
     "build:types": "pnpm tsc --build tsconfig-build.json && ./build-settings/check-type-definitions.ts",
     "build:storybook": "storybook build --stats-json",
     "clean": "rm -rf packages/wonder-blocks-*/dist && rm -rf packages/wonder-blocks-*/node_modules && rm -f packages/*/*.tsbuildinfo && rm -f utils/tsconfig.tsbuildinfo",
@@ -26,7 +27,7 @@
     "publish:check": "pnpm run lint:pkg-json && ./utils/publish/pre-publish-check-ci.ts",
     "publish:ci": "pnpm run publish:check && git diff --stat --exit-code HEAD && pnpm build && pnpm build:types && pnpm changeset publish",
     "start": "pnpm start:storybook",
-    "start:storybook": "storybook dev -p 6061",
+    "start:storybook": "pnpm dev & storybook dev -p 6061",
     "test:common": "pnpm run build && pnpm run lint && pnpm run typecheck && pnpm run alex",
     "test:coverage": "pnpm run test:common && pnpm run jest --coverage",
     "test:storybook": "vitest --project=storybook",
@@ -52,6 +53,9 @@
     "@khanacademy/eslint-config": "^5.2.0",
     "@khanacademy/eslint-plugin": "^3.1.1",
     "@khanacademy/wonder-stuff-testing": "^3.0.5",
+    "@khanacademy/wonder-blocks-link": "workspace:*",
+    "@khanacademy/wonder-blocks-tokens": "workspace:*",
+    "@khanacademy/wonder-blocks-theming": "workspace:*",
     "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-node-resolve": "^16.0.0",
     "@storybook/addon-a11y": "^8.5.2",
