+++ packages/wonder-blocks-theming/src/index.ts
@@ -10,3 +10,4 @@ export {
 export {type ThemedStylesFn, type SupportedThemes, type Themes} from "./types";
 export {default as useStyles} from "./hooks/use-styles";
 export {ThemeSwitcherContext} from "./utils/theme-switcher-context";
+export {ThemeSwitcher} from "./components/theme-switcher";
