+++ packages/wonder-blocks-theming/tsconfig-build.json
@@ -7,6 +7,5 @@
     },
     "references": [
         {"path": "../wonder-blocks-core/tsconfig-build.json"},
-        {"path": "../wonder-blocks-tokens/tsconfig-build.json"},
     ]
 }
