+++ packages/wonder-blocks-tokens/package.json
@@ -2,10 +2,19 @@
   "name": "@khanacademy/wonder-blocks-tokens",
   "version": "5.2.0",
   "description": "Core primitive design tokens for Web Wonder Blocks",
+  "exports": {
+    ".": {
+      "import": "./dist/es/index.js",
+      "require": "./dist/index.js"
+    },
+    "./styles.css": "./dist/css/index.css"
+  },
   "main": "dist/index.js",
   "module": "dist/es/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
+    "build": "node -r @swc-node/register ./src/build/generate-css-variables.ts",
+    "dev": "node --watch -r @swc-node/register ./src/build/generate-css-variables.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
     "prepublishOnly": "../../utils/publish/package-pre-publish-check.sh"
   },
@@ -14,7 +23,9 @@
   "publishConfig": {
     "access": "public"
   },
-  "dependencies": {},
+  "dependencies": {
+    "@khanacademy/wonder-blocks-theming": "workspace:*"
+  },
   "devDependencies": {
     "@khanacademy/wb-dev-build-settings": "workspace:*"
   }
