+++ packages/wonder-blocks-tokens/src/util/constants.ts
@@ -1,0 +1,10 @@
+/**
+ * The prefix for all CSS variables that are semantic colors.
+ *
+ * NOTE: The prefix uses the following structure:
+ * --wb-<namespace>-<token_category>-
+ *
+ * - namespace: The design token namespace (`s` for semantic tokens).
+ * - token_category: The category of the design token (e.g. `color`, `size`).
+ */
+export const CSS_VAR_COLOR_PREFIX = "--wb-s-color-";
