+++ packages/wonder-blocks-tokens/src/util/types.ts
@@ -1,0 +1,3 @@
+export type RecursivePartial<T> = {
+    [P in keyof T]?: RecursivePartial<T[P]> | string | number | boolean;
+};
