+++ packages/wonder-blocks-tokens/tsconfig-build.json
@@ -7,5 +7,7 @@
         "outDir": "./dist",
         "rootDir": "src"
     },
-    "references": []
+    "references": [
+        {"path": "../wonder-blocks-theming/tsconfig-build.json"},
+    ]
 }
